refactor(prediction-markets): decompose the DO into service modules
Split the 2127-line PredictionMarketsDO God-object into a 242-line thin RPC
facade plus eight standalone-function service modules under src/services/
(context, shared, read/wallet/market/betting/settlement/governance). Each RPC
method now delegates to a service function, passing a shared PmDeps { db }; the
cross-cutting helpers (creditWallet, logHistory, buildMarketDetail, sumStakes,
readActiveConfig, consumeRateBudget, mappers) live in services/shared.ts.

Pure mechanical move: method bodies are byte-identical except for threading the
db/deps parameter through the now-standalone helpers (verified by diffing every
money-critical and write method against the prior version). Typecheck, 74 unit
tests, and the worker build all pass; the public PredictionMarkets RPC surface
is unchanged.
